requirements.txt
PyYAML
watchdog
GitPython
PyGithub
tk
